사용자별 Role 추가

+++ auth-service/src/main/java/com/yangyag/msa/auth/factory/command/UserDeleteCommand.java
@@ -2,6 +2,7 @@ package com.yangyag.msa.auth.factory.command;
 
 import com.yangyag.msa.auth.model.dto.UserDeleteRequest;
 import com.yangyag.msa.auth.model.dto.UserUpdateRequest;
+import com.yangyag.msa.auth.model.entity.Role;
 import com.yangyag.msa.auth.model.entity.User;
 import com.yangyag.msa.auth.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -20,9 +21,14 @@ public class UserDeleteCommand implements Command<Boolean> {
     }
 
     private Boolean deleteUser(UserDeleteRequest request) {
+        if (request.getRole() != Role.ADMIN) {
+            throw new IllegalArgumentException("관리자만 사용자를 삭제할 수 있습니다.");
+        }
+
         User user = userRepository.findByUserId(request.getUserId())
                 .orElseThrow(() ->
                         new EntityNotFoundException(String.format("%s 값을 찾을 수 없습니다.", request.getUserId())));
+
         userRepository.delete(user);
         return true;
     }
